pyproject.toml
[project]
name = "subsquid-eth-archive"
description = "Subsquid stage 2 ingester and worker"
version = "0.1.0"
authors = [{name = "Subsquid Labs GmbH", email = "contact@subsquid.io"}]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "httpx == 0.*",
    "pyarrow == 10.*",
    "s3fs >= 2021.8.0",
]

[project.optional-dependencies]
eth-ingest = [
    "pycryptodome>=3.18.0",
    "prometheus-client==0.17.*",
]

http-worker = [
    "duckdb == 0.8.1",
    "falcon == 3.*",
    "marshmallow == 3.*",
    "sentry-sdk == 1.28.*",
    "uvicorn == 0.*",
    "psutil>=5.9.5",
]

p2p-worker = [
    "duckdb == 0.8.1",
    "grpcio == 1.*",
    "marshmallow == 3.*",
    "protobuf == 4.*",
    "sentry-sdk == 1.28.*",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "locust>=2.15.1",
]
